+++ sdk/entra/Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Removing 'TeantId' from the AuthenticationEventTrigger Attribute API configuration
 - Adding 'AuthorizedPartyAppId', 'AuthorityUrl' to the AuthenticationEventTrigger Attribute API configuration
 - Adding 'AuthenticationEvents__ShowPIIDataInLogsKey` (defaults to true) to configuration manager for logging PIIData in logs
+- Updated the Failed and Completed methods to not be async
 
 ### Bugs Fixed